refactor: 수동 로또 생성기 인터페이스 추가
