+++ src/main/java/domain/LottoGenerator.java
@@ -1,0 +1,5 @@
+package domain;
+
+public interface LottoGenerator {
+    Lotto generateLotto();
+}
