--- src/main/java/domain/LottoNumbersGenerator.java
@@ -1,7 +1,0 @@
-package domain;
-
-import java.util.Set;
-
-public interface LottoNumbersGenerator {
-    Set<LottoNumber> generateNumbers();
-}
